chore(physical): coding physical planner and adjust binder level in the transformer for operator code.

+++ src/include/common/type.h
@@ -6,7 +6,7 @@
 
 namespace YourSQL {
 
-    typedef uint64_t entry_id;
+    typedef size_t entry_id;
 
     typedef uint64_t page_id_t;
